+++ README.md
@@ -13,7 +13,7 @@
 
 
 - 스스로 생각하기에
- - [ ] 파일들을 폴더로 구분하기
+ - [x] 파일들을 폴더로 구분하기
  - [ ] 더욱 mvc 패턴처럼 해보기
  - [ ] 상수화
