+++ src/main/kotlin/lotto/model/LottoMaker.kt
@@ -2,11 +2,11 @@ package lotto.model
 
 import lotto.controller.validation.LOTTO_AMOUNT
 import lotto.controller.validation.RandomNumber
-import lotto.view.Output
+import lotto.view.OutputView
 
 class LottoMaker {
     private val randomNumber = RandomNumber()
-    private val output = Output()
+    private val outputView = OutputView()
 
     private fun getLottoCount(inputAmount : Int): Int {
         return inputAmount.div(LOTTO_AMOUNT)
@@ -16,7 +16,7 @@ class LottoMaker {
         val lottoCount = getLottoCount(inputAmount)
         val lottoNumbers = randomNumber.randomLottos(lottoCount)
 
-        output.printLottoNumbers(lottoNumbers)
+        outputView.printLottoNumbers(lottoNumbers)
         return lottoNumbers.map { Lotto(it) }
     }
 }
